+++ benches/distances_unweighted/graph_1.rs
@@ -26,7 +26,7 @@ fn dijkstra(bencher: Bencher<'_, '_>) {
     ];
 
     bencher.bench(|| {
-        let _ = dijkstra::min_distances_single_source(&graph, 0);
+        let _ = dijkstra::distances_single_source(&graph, 0);
     });
 }
 
@@ -44,6 +44,6 @@ fn bfs(bencher: Bencher<'_, '_>) {
     ];
 
     bencher.bench(|| {
-        let _ = bfs::min_distances_single_source(&graph, 0);
+        let _ = bfs::distances_single_source(&graph, 0);
     });
 }