+++ benches/distances_weighted/shortest_path_1.rs
@@ -1,4 +1,4 @@
-//! Benchmark [`graaf::algo::dijkstra::min_distances`] with different graph
+//! Benchmark [`graaf::algo::dijkstra::distances`] with different graph
 //! representations.
 fn main() {
     divan::main();
@@ -20,7 +20,7 @@ mod dijkstra {
         },
         alloc::collections::BinaryHeap,
         divan::Bencher,
-        graaf::algo::dijkstra::min_distances,
+        graaf::algo::dijkstra::distances,
         std::collections::HashSet,
     };
 
@@ -31,7 +31,7 @@ mod dijkstra {
         let mut heap = BinaryHeap::from(HEAP);
 
         bencher.bench_local(|| {
-            min_distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
+            distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
 
             dist
         });
@@ -50,7 +50,7 @@ mod dijkstra {
         let mut heap = BinaryHeap::from(HEAP);
 
         bencher.bench_local(|| {
-            min_distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
+            distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
 
             dist
         });
@@ -63,7 +63,7 @@ mod dijkstra {
         let mut heap = BinaryHeap::from(HEAP);
 
         bencher.bench_local(|| {
-            min_distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
+            distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
 
             dist
         });
@@ -82,7 +82,7 @@ mod dijkstra {
         let mut heap = BinaryHeap::from(HEAP);
 
         bencher.bench_local(|| {
-            min_distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
+            distances(&graph, |acc, w| acc + w, &mut dist, &mut heap);
 
             dist
         });